added gameboy

+++ src/app/login/login.js
@@ -8,7 +8,7 @@ angular.module('app.login', [])
         });
     })
 
-    .controller('LoginCtrl', function ($scope, $state, localStorageService, $rootScope) {
+    .controller('LoginCtrl', function ($scope, $state, localStorageService, $rootScope, $http, $log) {
         $scope.isFormTouched = false;
         var user = localStorageService.get('user') || {};
         var toCheckout = localStorageService.get('fromCheckout');
@@ -38,21 +38,30 @@ angular.module('app.login', [])
                 toastr.warning('<img src="assets/img/Epic_Mass_Facepalm.gif"/>');
                 toastr.warning('You had one Job!');
             } else {
-                toastr.success('Einloggen war erfolgreich.');
-                var role;
-                $scope.data.userName == 'daniel' ? role = 'admin' : role = 'user';
-                var user = {
-                    userName: $scope.data.userName,
-                    role: role
-                };
-                localStorageService.set('user', user);
-                $rootScope.$emit('login');
-                localStorageService.remove('checkout');
-                if (toCheckout) {
-                    $state.go('checkout');
-                } else {
-                    $state.go('home');
-                }
+                $http.get('http://localhost:8080/rest/user/' + $scope.data.userName)
+                    .then(function (response) {
+                        if (response.data.password === $scope.data.password) {
+                            toastr.success('Login war erfolgreich.');
+                            var role = response.Rolle;
+                            var user = {
+                                userName: $scope.data.userName,
+                                role: role
+                            };
+                            localStorageService.set('user', user);
+                            $rootScope.$emit('login');
+                            localStorageService.remove('checkout');
+                            if (toCheckout) {
+                                $state.go('checkout');
+                            } else {
+                                $state.go('home');
+                            }
+                        } else {
+                            toastr.warning('Benutzername und Passwort stimmen nicht überein!');
+                        }
+                    }, function (error) {
+                        $log.error(error);
+                        toastr.error('Fehler bei der Verbindung! Status ' + error.status);
+                    });
             }
         }
     });
